+++ src/core/conditions/isCondition.go
@@ -1,7 +1,7 @@
 package conditions;
 
 func IsCondition(expression string) bool {
-	var r bool = true
+	var r bool = false
 
 	return r
 }