+++ src/func/instancifyFunctions.go
@@ -1,19 +1,26 @@
 package IDENT
 
 import (
+	"fmt"
 	"memory/src/args"
 	"memory/src/core"
+	"os"
 )
 
-func instancyFunction(item core.Item) core.Function {
+func InstancyFunction(item core.Item) core.Function {
 	var args []core.Arg = args.Split(item)
 	var correctedSub core.Function
 
 	if item.SubFunction != nil {
-		correctedSub = instancyFunction(*item.SubFunction)
+		correctedSub = InstancyFunction(*item.SubFunction)
 	}
 
 	var availableFunctions map[string]core.Function = GetAvailableFunctions()
+
+	if !Contains(item.Name, GetAvailableFunctionsNames()) {
+		fmt.Printf("[73402] - At line %d - %s is not valid.\n", item.Line, item.Name)
+		os.Exit(1)
+	}
 
 	return core.Function{
 		Name: item.Name,
@@ -22,4 +29,13 @@ func instancyFunction(item core.Item) core.Function {
 		SubFunction: &correctedSub,
 		Execute: availableFunctions[item.Name].Execute,
 	}
+}
+
+func Contains(t string, l []string) bool {
+	for i := range l {
+		if l[i] == t {
+			return true
+		}
+	}
+	return false
 }