+++ src/func/mappedFunctions.go
@@ -5,8 +5,17 @@ import (
 	"memory/src/core/functions/console"
 )
 
+var names []string = []string{
+	"$print",
+}
+var functions map[string]core.Function = map[string]core.Function{
+	"$print": console.Print_D(),
+}
+
 func GetAvailableFunctions() map[string]core.Function {
-	return map[string]core.Function{
-		"$print": console.Print_D(),
-	}
+	return functions
+}
+
+func GetAvailableFunctionsNames() []string {
+	return names
 }